Manage certificate storage in util
